+++ xmake/modules/core/tools/gcc.lua
@@ -345,21 +345,35 @@ function nf_linkgroup(self, linkgroup, opt)
     end
     local flags = {}
     local extra = opt.extra
-    if extra and not self:is_plat("macosx") then
-        local group = extra.group
+    if extra and not self:is_plat("macosx", "windows", "mingw") then
+        local as_needed = extra.as_needed
         local whole = extra.whole
-        if group and whole then
-            -- https://github.com/xmake-io/xmake/issues/4308
-            table.join2(flags, "-Wl,--whole-archive", "-Wl,--start-group", linkflags, "-Wl,--end-group", "-Wl,--no-whole-archive")
-        elseif group then
-            table.join2(flags, "-Wl,--start-group", linkflags, "-Wl,--end-group")
-        elseif whole then
-            table.join2(flags, "-Wl,--whole-archive", linkflags, "-Wl,--no-whole-archive")
-        end
+        local group = extra.group
         local static = extra.static
+        local prefix_flags = {}
+        local suffix_flags = {}
         if static then
-            table.join2(flags, "-Wl,-Bstatic", linkflags, "-Wl,-Bdynamic")
+            table.insert(prefix_flags, "-Wl,-Bstatic")
+            table.insert(suffix_flags, 1, "-Wl,-Bdynamic")
         end
+        -- https://github.com/xmake-io/xmake/issues/5621
+        -- https://github.com/xmake-io/xmake/pull/5631
+        if as_needed then
+            table.insert(prefix_flags, "-Wl,--as-needed")
+            table.insert(suffix_flags, 1, "-Wl,--no-as-needed")
+        elseif as_needed == false then
+            table.insert(prefix_flags, "-Wl,--no-as-needed")
+            table.insert(suffix_flags, 1, "-Wl,--as-needed")
+        end
+        if whole then
+            table.insert(prefix_flags, "-Wl,--whole-archive")
+            table.insert(suffix_flags, 1, "-Wl,--no-whole-archive")
+        end
+        if group then
+            table.insert(prefix_flags, "-Wl,--start-group")
+            table.insert(suffix_flags, 1, "-Wl,--end-group")
+        end
+        table.join2(flags, prefix_flags, linkflags, suffix_flags)
     end
     if #flags == 0 then
         flags = linkflags